requirements.txt
# Core framework
crewai==0.130.0
crewai-tools==0.47.1

# API server
fastapi==0.110.3
uvicorn[standard]>=0.29.0

# PDF processing — used by PyPDFLoader in tools.py
langchain-community>=0.2.0
pypdf>=4.0.0

# LLM / AI
openai==1.30.5

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic==1.10.13
python-multipart>=0.0.9